cloud-functions/consume_firestore/requirements.txt
google-cloud-pubsub
google-cloud-core
firebase_admin